refactor(server): split OrderService into CheckoutService and ReportService
